Initial documentation with main structural elements.
The documentation has been brought over from the HTML specification
created from the Impression version (issue 3) to make a basic document
that has structural elements present (section, subsection, category).

The structure and style does not fit with the documentation style. It
needs reworking later, but the first things to do are to try to get it
into the PRM-in-XML format.

The MBufManager has been split off, mostly (although a section remains),
so that it can be dealt with separately.