Updated the SWI and services to use PRM-in-XML format.
The SWIs and services are now defined in the standard format to
make them easier to use. This means they're more verbose now, because
the reason codes have been split out into separate definitions in some
cases. But this allows them to be more easily expanded upon.

The name of the MbufManager has also been corrected.