Restructure the statistics to use swi and service definitions.
The statistics document has been updated to have more of the
SWI and service definitions documented, and to change some of the
inline definitions into structures. This should make it easier to
move them around.

Some of the moves have already been done, with the operation moved
into an earlier section rather than being described in the SWI,
following the principle of describing the technical details and
leaving the SWI to deal with the implementations.