Restructure into Technical Details and the usage information.
The restructured document reads a lot better - it's a 'here's what
the concepts are and what you'd need to do to use them', followed
by the SWI and services that you'd actually use.

Also made a small change to the stylesheet to make the 'Name' heading
appear by default on the value-table and offset-table, in the same
way as they do on the bitfield-table.